fix: Model exposes set of creatable Element classes
The Model exposes a set of the classes supported by ElementFactories provided to the model in the Lookup it was created with. This allows the burden of determining how to display this information to be placed purely on the UI, but allows the UI to know what the Model supports.